refactor: hooks 디렉토리 구조 정리 및 recommendation 훅 분리
