+++ pages/admin/recommendation/[id]/edit.js
@@ -1,7 +1,7 @@
 import { useRouter } from "next/router";
 import { useEffect, useState } from "react";
-import { useRecommendationDetail } from "@/hooks/useRecommendationDetail";
-import { useUpdateRecommendation } from "@/hooks/useUpdateRecommendation";
+import { useRecommendationDetail } from "@/hooks/recommendation/useRecommendationDetail";
+import { useUpdateRecommendation } from "@/hooks/recommendation/useUpdateRecommendation";
 import RecommendationForm from "@/components/recommendation/RecommendationForm";
 import { uploadImage } from "@/lib/firebase/uploadImage";
